Added files from Aaron.
"Here are the files its still not 100% but it will work, it needs to
have the stuff in the index class fixed still in order to change what
is printed out on for login, my account, logout. I am working on the
error checking now but i think i may need to get the stuff on the index
page working first. Let me know if you have any problems. Also only one
log in will work unless you create a new account cause i changed the
format to the hashed password. test81@test.com - test (is the passwd)"